backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
attrs==24.3.0
autopep8==2.3.1
azure-core==1.32.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.8
cryptography==44.0.0
distro==1.9.0
exceptiongroup==1.2.2
flake8==7.1.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
ics==0.7.2
idna==3.10
iniconfig==2.0.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
msal==1.31.1
msal-extensions==1.2.0
openai==1.61.1
packaging==24.2
pluggy==1.5.0
portalocker==2.10.1
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.3
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.11.0
SQLAlchemy==2.0.36
TatSu==5.13.1
tomli==2.1.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
